Android 17.0.0 release 1

* tag 'android-17.0.0_r1': (296 commits)
  Fix permanent BAL grant via onServiceConnected race
  Fix permanent BAL grant via onServiceConnected race
  Fix permanent BAL grant via onServiceConnected race
  Nfc: Fix requiresScreenOn check for HCE
  [Nfc] Skip flaky HCE multi-device tests
  Call unbindService() when bindServiceAsUser() fails
  [Nfc CTS] Prevent PN532 timeouts and increase TagLostException timeout
  [Nfc CTS] Add test for NDEF write verification
  Import translations. DO NOT MERGE ANYWHERE
  Separate the tap-to-x trigger flow post Android C
  [nfc] Skip testRegisterAndUnregisterGestureExchangeCallbacks test if the Vendor API Level is less than 202504
  onDeactiviated is not triggered when updating discovery state
  Remove observe mode always on
  Remove observe mode always on tests
  Add CTS to restrict PERFORM_GESTURE_EXCHANGE declaration
  Query Set/Get Configs for Type-F only if Type-F is supported
  Support NFC tap-to-x in home screen
  nfc(cts): Skip the tests if reader mode is not supported
  Add provision to ignore proprietary intf activated ntf
  Fallback to route DH only if default route is not DH
  ...

+++ libnfc-nci/src/nfa/dm/nfa_dm_api.cc
@@ -1237,7 +1237,7 @@ tNFA_STATUS NFA_SendVsCommand(uint8_t oid, uint8_t cmd_params_len,
 **                  NFA_STATUS_FAILED otherwise
 **
 *******************************************************************************/
-tNFA_STATUS NFA_SendRawVsCommand(uint8_t cmd_params_len, uint8_t* p_cmd_params,
+tNFA_STATUS NFA_SendRawVsCommand(uint16_t cmd_params_len, uint8_t* p_cmd_params,
                                  tNFA_VSC_CBACK* p_cback) {
   if (cmd_params_len == 0x00 || p_cmd_params == nullptr || p_cback == nullptr) {
     return NFA_STATUS_INVALID_PARAM;
